+++ oci/blockchain_blockchain_platform_resource.go
@@ -291,6 +291,10 @@ func BlockchainBlockchainPlatformResource() *schema.Resource {
 				Type:     schema.TypeString,
 				Computed: true,
 			},
+			"platform_shape_type": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
 			"service_endpoint": {
 				Type:     schema.TypeString,
 				Computed: true,
@@ -760,6 +764,8 @@ func (s *BlockchainBlockchainPlatformResourceCrud) SetData() error {
 	}
 
 	s.D.Set("platform_role", s.Res.PlatformRole)
+
+	s.D.Set("platform_shape_type", s.Res.PlatformShapeType)
 
 	if s.Res.Replicas != nil {
 		s.D.Set("replicas", []interface{}{ReplicaDetailsToMap(s.Res.Replicas)})